move functionality into process_operations
